boundaries/pyproject.toml
[project]
name = "boundaries"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "geopandas",
    "pyogrio",
    "shapely",
    "pyproj",
    "psycopg2-binary",
    "requests",
]
